requirements.txt
# Core Libraries
Flask==3.0.3
google-generativeai==0.3.2
twilio==9.0.1
pyspellchecker==0.8.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
python-dotenv==1.0.1
gunicorn==22.0.0